+++ module/core/variadic_from/tests/variadic_from_tests.rs
@@ -3,8 +3,6 @@
 #[ allow( unused_imports ) ]
 use variadic_from as the_module;
 #[ allow( unused_imports ) ]
-use variadic_from;
-#[ allow( unused_imports ) ]
 use test_tools::exposed::*;
 
 #[ cfg( feature = "enabled" ) ]